♻️  익스텐션 정리, 함수 위치 정리
